implement edge classification of directed graph
being either:
- tree edges
- forward edges
- cross edges
- back edges

We take a simple list structure representing a graph, where the car of each
sublist is a node id, and the cdr is a list of edges to successors, represented
as conses. The first sublist (B0) is interpreted as the entry node:

(defparameter maxcol-graph
  '((0 ((0 . 1) (0 . 5))) ;; B0
    (1 ((1 . 2) (1 . 4))) ;; B1
    (2 ((2 . 3) (2 . 6))) ;; B2
    (3 ((3 . 2) (3 . 4))) ;; B3
    (4 ((4 . 1) (4 . 5))) ;; B4
    (5 ())                ;; B5
    (6 ((6 . 3)))))       ;; B6

This structure is transformed into node classes, where references to successor
nodes are stored in the nodes. The nodes can be subclassed to add functionality
for different types of analysis.