transform a simple source-level function into an IR flow graph
Using the datastructures from the previous patches, we can now compile the
following snippet into an IR flow graph:

(defun add-it (a b)
  (add a b))

We walk the expressions and:
- we register that we are compiling a function, with name add-it
- so we start a new IR flow graph, inialize the entry and exit basic blocks and
link them together. We set the entry basic block to being the currently active
one.
- we also create a recursive environment. The function environment has the
global environment as a parent to include in the variable lookup
- we record the function parameters in the base function environment and we
create IR temps for them
- we then parse the add function application, and recognize that this is a
primitive function
- we look up the function arguments in the environment to retrieve the
temporaries
- we create an IR op, and create a new temporary for the op result.
- we append the IR op to the currently active basic block (the entry block).
- the function walk code returns the result temporary which we currently don't
do anything with.