add poison value/type to IR
Let's see if this is a good idea, but I'd like to be able to track if we
shouldn't really use certain values.

The particular example I had in mind is an if expression with only a true
branch. In lisp every s-expression returns a value. So you need to decide how
you want to merge the two branches of an if expression. What value do you want
to return for the false branch? You could return a suitable null value for the
type, you could return some universal null value and have actual true and false
values. Just for now I decided to not rely on implicit returns from a false
branch. If you want to return a certain value, do so explicitly.

So we return a poison value, which if used later on will trigger an error. So an
if statement with only a true value is expected to be executed only for
side-effects. Admittedly this is an ugly solution, but I thought it'd be a nice
way to play around with the concept of a poison value.

changes:
- to materialize an IR poison value, use the `ldp` or `load poison` operator
- `if` expressions with only a true branch will emit a poison value for the
  false branch.
- unrelated small fixup, renamed Lasp `get pointer of array subscipt` intrinsic
  `aloc` to `aptr`, as `aloc` sounds too much like `alloc`, which I wanted to
   use for allocations