hook up basic blocks successors and predecessors
Create functions for emitting branches and jumps. As part of this function, the
targets of these instructions are the successors and the current block to emit
to is the predecessor. Tie these blocks together.

And now that we have a directed graph, we can classify it, and use the node
classification to generate a reverse post-order list of the bbs to serialize
them in a sensible order.