register for temps if they're either expression or variable ones
essentially expression temporaries are the result of 'simple' instructions that
don't change because of assign, load or store, mem changes. They have the nice
property that we only need to recompute them if a temp that was reassigned
because of load/assignment/etc has changed, if certain properties hold.

This will be exploited in various places in the compiler (ex. computing
liveness)