initial scaffolding for parsing IR text representation to IR
Want to byte the bullet quickly on this as preparation for the fact that I don't
want to maintain a Lasp parsing and pure IR implementation of
creating/maintaining the temp table/BB CFG, etc. It would help if we can parse
pure IR, so we have an easy way to test and we no there's no wonky dependencies
anywhere or chance of them growing.